refactor(file-browser): extract all sub-components and add editor polish
- Extract FileEditorModal from GameAssetsBrowserView with:
  - Cmd/Ctrl+S save shortcut
  - Esc close with unsaved-changes confirm() guard
  - Live dirty-state indicator (• dot in header)
  - Disabled save button when no changes
  - maxLength 10M chars (~10 MB server limit)
  - Character counter in footer

- Rewrite GameAssetsBrowserView (1632 → ~580 lines) to import all
  sub-components from their own files:
  Toolbar, AssetGrid, FolderTree, SearchInput, ActionDropdown,
  ImagePreviewModal, AudioPlayerModal, ImageInfoPopover, FileEditorModal

- Extract client constants (CATEGORY_ICONS, DEFAULT_DESCRIPTIONS) to
  constants.tsx and file-type helpers (isImage, isAudio, isEditableText,
  FileIcon, countItems) to utils.tsx

- Extract human-readable format helpers (formatBytes, formatDate) to
  lib/format.ts for reuse across components

- Export game-asset constants (IMAGE_EXTS, AUDIO_EXTS, TEXT_EXTS,
  AUDIO_MIME_MAP) from @marinara-engine/shared for client+server reuse

- Remove inline duplicate component definitions, use improved extracted
  versions with proper outside-click and scroll-close handling
