+++ packages/client/src/lib/format.ts
@@ -1,0 +1,19 @@
+// ──────────────────────────────────────────────
+// Human-readable formatting helpers
+// ──────────────────────────────────────────────
+
+/** Format a byte count as a human-readable string (e.g. "1.5 MB"). */
+export function formatBytes(bytes?: number): string {
+  if (bytes == null) return "—";
+  if (bytes < 1024) return `${bytes} B`;
+  if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
+  if (bytes < 1024 * 1024 * 1024) return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
+  return `${(bytes / (1024 * 1024 * 1024)).toFixed(1)} GB`;
+}
+
+/** Format an ISO timestamp as a short absolute date (e.g. "May 10, 2026"). */
+export function formatDate(iso?: string): string {
+  if (!iso) return "—";
+  const d = new Date(iso);
+  return d.toLocaleDateString(undefined, { year: "numeric", month: "short", day: "numeric" });
+}
