+++ packages/shared/src/index.ts
@@ -48,6 +48,7 @@ export * from "./constants/agent-prompts.js";
 export * from "./constants/impersonate.js";
 export * from "./constants/image-generation-defaults.js";
 export * from "./constants/security.js";
+export * from "./constants/game-assets.js";
 
 // Utils
 export * from "./utils/macro-engine.js";
